Disable include toggle for required fields when recommended unchecked
